+++ crypto/src/test/kotlin/org/kethereum/crypto/TheSignatures.kt
@@ -13,7 +13,7 @@ class TheSignatures {
         val signatureData = SignatureData(
                 "0x0031f6d21dec448a213585a4a41a28ef3d4337548aa34734478b563036163786".hexToBigInteger(),
                 "0x2ff816ee6bbb82719e983ecd8a33a4b45d32a4b58377ef1381163d75eedc900b".hexToBigInteger(),
-                27.toByte()
+                27.toBigInteger()
         )
 
         assertThat(signatureData.toHex()).isEqualTo("0031f6d21dec448a213585a4a41a28ef3d4337548aa34734478b5630361637862ff816ee6bbb82719e983ecd8a33a4b45d32a4b58377ef1381163d75eedc900b1b")
