+++ etherscan/src/main/kotlin/org/walleth/kethereum/etherscan/EtherScan.kt
@@ -1,19 +1,20 @@
 package org.walleth.kethereum.etherscan
 
 import org.kethereum.model.ChainDefinition
+import java.math.BigInteger
 
 
 private val ETHERSCAN_PREFIX_MAP = mapOf(
-        1L to null,
-        3L to "ropsten",
-        4L to "rinkeby",
-        5L to "goerli",
-        42L to "kovan"
+        BigInteger.valueOf(1L) to null,
+        BigInteger.valueOf(3L) to "ropsten",
+        BigInteger.valueOf(4L) to "rinkeby",
+        BigInteger.valueOf(5L) to "goerli",
+        BigInteger.valueOf(42L) to "kovan"
 )
 
 val ALL_ETHERSCAN_SUPPORTED_NETWORKS = ETHERSCAN_PREFIX_MAP.map { it.key }.toSet()
 
-private fun getPrefix(chain: ChainDefinition) =  ETHERSCAN_PREFIX_MAP[chain.id.value]
+private fun getPrefix(chain: ChainDefinition) = ETHERSCAN_PREFIX_MAP[chain.id.value]
 
 private fun getAPIPrefix(chain: ChainDefinition) = getPrefix(chain)?.let { "-$it" } ?: ""
 fun getEtherScanAPIBaseURL(chain: ChainDefinition) = "https://api" + getAPIPrefix(chain) + ".etherscan.io"
